adding navbar and routes to browse charts
