+++ package.json
@@ -10,7 +10,9 @@
     "d3-geo-projection": "^2.0.0",
     "d3-scale-chromatic": "^1.0.0",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.2.0",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
